+++ src/index.ts
@@ -6,6 +6,7 @@ import replaceCustomValues, {
 export * as typeCastService from './typeCastService'
 export * as userService from './userService'
 export * as abnService from './abnService'
+export { validatePaymentAmount } from './paymentAmountValidation'
 export { getContentDisposition } from './fileUploadService'
 
 export { replaceCustomValues, getElementSubmissionValue }